refactor: cleanup helpers
