+++ src/codegen/actions/tables/templates/delete-item.ts
@@ -1,5 +1,5 @@
 import {ensureTableTemplate} from './ensure-table';
-import {objectToString} from './mappers';
+import {objectToString} from './helpers';
 
 export interface DeleteItemTplInput {
   readonly key: Record<string, string>;
