--- src/codegen/actions/tables/templates/mappers.ts
@@ -1,4 +1,0 @@
-/** Converts a compile time object to a runtime object */
-export function objectToString(obj: Record<string, string>): string {
-  return `{${Object.entries(obj).map(([k, value]) => `${k}: ${value}`)}}`;
-}
