+++ src/codegen/actions/tables/templates/read-item.ts
@@ -1,5 +1,5 @@
 import {ensureTableTemplate} from './ensure-table';
-import {objectToString} from './mappers';
+import {objectToString} from './helpers';
 
 export interface ReadItemTplInput {
   readonly consistent: boolean;
